Moves monitors module componenets to new componenets directory
